rag-service/requirements.txt
fastapi
uvicorn
mem0ai
python-dotenv
qdrant-client
